refactor: AzureResourceGroup into ResourceEntity class and interface
