--- src/prod/AzureReaper.Functions/Interfaces/IAzureResourceGroup.cs
@@ -1,8 +1,0 @@
-namespace AzureReaper.Functions.Interfaces;
-
-public interface IAzureResourceGroup
-{
-    public bool GetSchedule();
-    public void CreateResource(string resourceId);
-    public void DeleteResource();
-}
