+++ src/prod/AzureReaper.Functions/Interfaces/IResourceEntity.cs
@@ -1,0 +1,11 @@
+using System.Threading.Tasks;
+
+namespace AzureReaper.Functions.Interfaces;
+
+public interface IResourceEntity
+{
+    public bool GetSchedule();
+    public Task CreateResource(string resourceId);
+    public void DeleteResource();
+    public Task CheckReaperTag();
+}
